hilt update

+++ app/src/main/java/com/example/alarmmanager/UserData.kt
@@ -3,5 +3,5 @@ package com.example.alarmmanager
 data class UserData(
     val firstName: String?,
     val lastName: String?,
-    val time: Int?
+    val time: Long?
 )
